Added extra example, added descriptions to other examples

+++ matrix.cs
@@ -3,6 +3,18 @@
  * Copyright: (c) 2018 Gerrard de Jonge <development.gdj@netvisit.nl>
  * License:   GPLv2 (see file COPYING)
  *
+ *
+ * Description
+ *            Prints a matrix for a given height and width
+ *
+ * Example
+ *            Enter number of rows: 2
+ *            Enter number of columns: 3
+ *            +--+--+--+
+ *            |  |  |  |
+ *            +--+--+--+
+ *            |  |  |  |
+ *            +--+--+--+
  */
  
 
@@ -15,9 +27,9 @@ public class Matrix
 	
 	public void GetEntries() {
 		Console.Write("Enter number of rows: ");
-		rows = UInt16.Parse(Console.ReadLine());
+		rows = Int32.Parse(Console.ReadLine());
 		Console.Write("Enter number of columns: ");
-		cols = UInt16.Parse(Console.ReadLine());
+		cols = Int32.Parse(Console.ReadLine());
 	}
 		
 	public void PrintColumns(String s) {
